Refactor fractal generation (#5)
Split apart fractal generation into separate parts. compute.go computes the number of iterations, it gets stored in the new Result struct, and then color.go contains functions for transforming that into a color spectrum.

+++ fractal/params.go
@@ -11,7 +11,6 @@ type Params struct {
   YMax        *big.Float
   Width       int
   Height      int
-  Iterations  uint8
+  Iterations  int
   Escape      *big.Float
 }
-
